feat: add in a working watchdong configuration
This watchdog configuration ensures that the interval is always around
half of the timeout of thw watchdog.

The underlying reason of making it half is to ensure that the interval
itself does not take far longer than the watchdog timeout which notifies
the necesary time for the timeout to occur. Making it too close to 15
(or more) will cause a boot loop.